+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -13,6 +13,8 @@ however, it has to be formatted properly to pass verification tests.
 ### Changed
 - Removed icons from action map list as these were always the same and the icon was placeholder
 
+### Added
+- Support for [Game rotation vector](https://developer.android.com/reference/android/hardware/Sensor#TYPE_GAME_ROTATION_VECTOR) sensor on Android
 
 ## [1.8.0-pre.1] - 2023-09-04